Revert "Fixed crash with FixedDiv(-2147483648,x)"
This reverts r2926, which has been found to cause desyncs.
See prboom-plus r3193.

+++ prboom2/src/m_fixed.h
@@ -142,16 +142,7 @@ inline static CONSTFUNC fixed_t FixedMul(fixed_t a, fixed_t b)
 #pragma warning( disable : 4035 )
 __inline static fixed_t FixedDiv(fixed_t a, fixed_t b)
 {
-    // e6y
-    // zg is a master of engineer science.
-    //
-    // Fixed crash with FixedDiv(-2147483648,x)
-    // Exception Number : EXCEPTION_INT_OVERFLOW(C0000095)
-    //
-    // Some other ports (Eternity, Chocolate) return wrong value instead of MAXINT.
-    // For example FixedDiv(-2147483648,-30576) should return INT_MAX instead of 307907126
-    // 307907126 is truncated correct int64 value: 4602874423 - 2^32 = 307907126
-    if ((unsigned)D_abs(a) >> 14 >= (unsigned)D_abs(b))
+    if (D_abs(a) >> 14 >= D_abs(b))
         return (a^b)<0 ? INT_MIN : INT_MAX;
     __asm
     {
@@ -176,8 +167,7 @@ __inline static fixed_t FixedDiv(fixed_t a, fixed_t b)
 inline
 static CONSTFUNC fixed_t FixedDiv(fixed_t a, fixed_t b)
 {
-  //e6y: zg is a master of engineer science
-  if ((unsigned)D_abs(a) >> 14 < (unsigned)D_abs(b))
+  if (D_abs(a) >> 14 < D_abs(b))
     {
       fixed_t result;
       asm (
@@ -200,7 +190,7 @@ static CONSTFUNC fixed_t FixedDiv(fixed_t a, fixed_t b)
 
 inline static CONSTFUNC fixed_t FixedDiv(fixed_t a, fixed_t b)
 {
-  return ((unsigned)D_abs(a)>>14) >= (unsigned)D_abs(b) ? ((a^b)>>31) ^ INT_MAX :
+  return (D_abs(a)>>14) >= D_abs(b) ? ((a^b)>>31) ^ INT_MAX :
     (fixed_t)(((int_64_t) a << FRACBITS) / b);
 }
